Added test changes for consistency with URLVoid service tests
This commit applies some changes used in the URLVoid project tests for
avoiding some code needed to create and connect a DXL client for use in
registering a service. This commit also includes a line to close the
mock HTTP server during shutdown - for which a warning message had
started appearing when the tests are run in Python 3.7.